requirements.txt
# Dependencies for the Mongolian bank exchange-rate API
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0

# HTTP requests
requests>=2.31.0

# Web scraping
beautifulsoup4>=4.12.0
playwright>=1.40.0
lxml>=4.9.0

# Code quality (dev only)
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pylint>=3.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
respx>=0.20.0